Implements simple rolling deploy logic.
Requires the latest ansible to use (you may need to remove and add it again if
you are in a virtualenv)

See `edxapp_rolling_example.yml`

A new module is necessary for ELB information since that is not
available in EC2 metadata (which is what ec2_facts fetches).

ec2_facts grabs the instance ids and other meta information
ec2_elb_facts fetches ELB information
elb_reg.py takes an elb and instance id and will either register or
deregister. (called as pre and post).

`Serial` should be set to the number of instances that you want to
operate on simultaneously, currently set to 1.