enabling 'private_ip_address' so ec2.py can find hosts in the VPC
This presumes that we have some way to actually *get* to hosts in the
VPC.  We still need to figure that out -- could be DNS with some fancy
routing, could be a statically generated/distributed ssh config file.
But we have to still do ssh forwarding through the Bastion to get to
these machines, even if ec2.py will now list them.