cloudformation: open tcp/9418 for git
We need to use the git protocol to access read-only git repos.  It's a
quirk that when these are referenced from pip files, they need to go
over a source control protocol.  I wish it could go over http/s or ssh,
but it can't, so port 9418 it is.

I had to open inbound and outbound 9418 to get this working, don't know
why.  Leaving both open for now, will debug later so we can close down.