Update the ruby role for issues found while testing jenkins role.
In certain states, ruby is installed but the env is not setup so it is importart
to always run the environment update task and the rehash task.

The rbenv install task prompts if the version of ruby is already installed. This
freezes the build so skip it if the version directory already exists.