Stanford 'certificates' stopgap branch
Substantially similar to the certs role developed (in paralell) that
exists on master, but this one uses pre-Croissant directory structure
and variable conventions. Modeled on the croissant-era notifier role.
Useful as a bridge to midterms, when Stanford can afford to port to the
new ways of doing things.

Includes updates to logging-related tasks for ansible and datadog. These
are ongoing problems for us at the moment.