DO NOT MERGE: Pure ansible provisioning and MITx style architecture This adds roles and a play for *mostly* building a stack using pure ansible. The roles build the MITx style stack with split AZ simple tiers.  There is also a config variable to have this build a non HA stack.  There is also a "load balancer" role that just rsyncs the static content from the first app server to the nginx roled server using whatever ssh key it can find.
Mostly this is just to show what is going on and isn't what I would
call public ready.  Particularly because it relies on two of my
PRs to the ansible repo.  One is just a module, but the other is an
enhancement to an existing module that lets you assign public IPs to
instances in a non-default VPC so that you don't need all the NAT
stuff.