runs the worker role before the edxapp role
Because the CMS uses the git hash of the edx-platform for the
static path it can be problematic if you update the edx-platform
repo for the worker AFTER you update the edx-platform repo
for edxapp.  When running the edxapp role in worker mode
rake gather_assets is skipped.

This is a workaround but we may want to consider using separate
repos for worker and edxapp for situations where both are installed
on the same server.