Make LC_ALL default to 'en_US.UTF-8'.
The LC_ALL environment variable is set to en_US in /etc/default/locale
in the precise32 vagrant box, see:

https://github.com/mitchellh/vagrant/issues/1188

When running the python unit tests in the devstack machine, several of the
tests fail with a UnicodeEncodeError when trying to create folders with names
that contain non-ascii characters.

Explicitly setting LC_ALL to en_US.UTF-8 before running the tests fixes this issue,
but it would be nice if the box was set up with the en_US.UTF-8 locale by default.