TE-414: packer jenkins_worker: use ansible 1.7.1 for testing build server.
The test_build_server role was failing on acceptance tests because of unicode
static assets being gathered in collectstatic for django. This was because of
this issue https://github.com/ansible/ansible/issues/6655, which has been
resolved in recent ansible versions. We must override the version that is
being used by this repo, and upgrade ansible at provisioning time for
running the test_build_server role.