Don't run `/edx/bin/update`  with `--tags deplay` by default.
Having `--tags deploy` be part of the ansible-playbook call
that `/edx/bin/update` makes is not intuitive.  It's broken
us a few times recently and I don't think the benefit it
provides is worth the cost and confusion.  I feel that anyone
who wants to use tags will know how to do it without using
`/edx/bin/update`.