Stop using the RABBITMQ_CLUSTERED var and tooling.
This method relied on the `hostvars` variable provided by ansible to always
be all the nodes in the rabbitmq cluster.  The hostvars var broke as a result
of the upgrade to ansible 1.9 but also even if it did work, this method
required that you always run rabbit plays on the whole cluster.  This is not
reasonable as it makes it very hard to heal a single node that may have fallen
over.

From here on, the RABBITMQ_CLUSTERED_HOSTS var should be used instead to
specify the rabbit hosts that are part of a cluster.