Update migrate.sh for Dogwood upgrades
Upgrading to Dogwood is complicated, because we have to make two
intermediate stops along the way: once at the end of Django 1.4 to
migrate the databases using South, then to the beginning of Django 1.8
to apply --fake-initial migrations.  Then we can upgrade to the Dogwood
release that we actually want.