Don't make this a serial play
We usually run this on 3 machines, so serial: 3 was equivalent to
ansible's default of "run everything in parallel" but this causes
problems if you ever run it on 4 like we do for prod envs.

In addition, this prevents run_once from working properly, and there are
a number of things we only one to do on one machine (like creating a
superuser).