Do a bit more error checking before asking for the replica set.
The new mongo play wants to be able to run this against an unconfigured
Mongo to find out "Do you even have a replica set?" which this can now
answer (the try/except portion).

In the case where mongo was started without a replSet configured (or
running getCmdLineOpts fails for an OperationFailure reason), we
return a dictionary where status is not included to allow tasks to
detect the failure.  Unfortunately, Ansible doesn't provide an easy way
to log back why you didn't get a replica set doc.