Make the module work without rs being configured
Initially, this module always tried to find and connect to a primary,
however, during replica set initialization, there is no primary so we
need to fall back to just connecting to the machine specified.

This is detected by catching the replSetGetStatus failure and falling
back to the bare get_client method which refactors the connection.