Refactor our logic into two functions and two loops
I believe the loops will be easier to debug in the long run.
Essentially, we have a shallow compare loop in compare_rs_config
and then a deep compare loop in compare_rs_members.

Both functions allow extra information in the config that comes back
from Mongo, since Mongo tracks things like votes: slaveDelay: etc that
may not be reflected in the configuration you pass in.