Refactor of role to handle initialization and reruns
Previously on this branch, it would attempt to create a superuser and
initialize the replica set after the local host exemption had passed
(once a mongod is started with a keyfile and the replicationSetName
 config file option or --replSet command line argument, you can't create
 users using the localhost exemption anymore ).

This necesarily complicates the logic and requires us to track if we're
initializing a replica set or not (with facts) and to template out
/etc/mongod.conf twice in the initialization case.

Further modifications to the replica set can be run against all the
members and will install mongo on new machines while updating the
replica set on the old machines.  With a further commit on this branch,
we can avoid needlessly restarting mongo on all machines in the cluster.

When we run this on an existing cluster, this stops and starts mongo for
all machines it is run against.  We can't rely on the notify: since that
happens at the end of the play and during initial cluster setup, we need
to restart and pick up the changed config.

Try to stick sensible tags on things

Ensure mongo restarts on the cluster key file change

Add a few playbook notes about common operations