Don't always stop mongo
When we run this on an existing cluster, this stops and starts mongo for
all machines it is run against.  We can't rely on the notify: since that
happens at the end of the play and during initial cluster setup, we need
to restart and pick up the changed config.

A later change to mongod.conf has a restart that is explicitly gated.