Freeze pip, setuptools, and virtualenv
Pip 8.0.0, released a few hours ago, tries to uninstall deb-installed
python modules that it wants to upgrade, and always fails when doing so.
The effect is that configuration/requirements.txt cannot be fulfilled,
notably ansible 1.9.3-edx, which requires an upgrade to pycrypto.

This forces the installation of pip to the latest one prior to 8.0.0,
7.1.2.

It also freezes setuptools and virtualenv, so that they should all match
the current versions required on
playbooks/roles/common_vars/defaults/main.yml:common_pip_pkgs.