Remove almost all references to syncdb
Django 1.8 removed the syncdb command so we can no longer run it, don't
imply that we are running it in documentation/messages, also don't
provide the scripts used on devstack/sandboxes to run it since they'll
never work anyway.

Additionally - remove yet another openid workaround which is unnecesary
since the Django 1.8 upgrade.