Run migrate on every DB in DATABASES for both lms and the cms
Does so by running the edxapp role just to generate config (as
controlled by check-migrations.sh's use of the --tag) and then
the tasks in the playbook are hardcoded to always run, regardless
of the tag.  Because the edxapp play has run, the tasks have
access to the various config variables containing DATABASES
and can tell what DBs are in use.

Skip the read replica