This registers over the microsites theming
This clobbers edxapp_theme_checkout set by the previous task, even when
the task is skipped because EDXAPP_COMPREHENSIVE_THEME_SOURCE_REPO is
unset.

edxapp_theme_checkout is used to tag the instance when the edxapp play
is run.

That tag is later used by our build infrastructure to ensure that
theme is consistent between builds.

If we want to also tag the comprehensive theme value, that's fine, but
it should not clobber the value needed by the other type of themes, and
should declare a different tag.