OPS-1385 role for mounting/formatting volumes
This implements a role that consumes a volume list of dicts and ensures
those are all mounted in the expected location on your system as well as
formatting any drives with the wrong fstype.

As a safety feature, it defaults to being unwilling to unmount drives
(since that might result in data loss or mismanagement) and there is a
flag you can -e to enable it.

It is assumed you will run this role passing along your volume list from
your main playbook and secure vars.