Flag rabbit upgrading
Adds the CHECK_RABBIT_VERSION flag (true by default). If true and the
rabbitmq version found differs from the version specified in
RABBITMQ_VERSION, the role fails. If false in the same scenario, the
role continues but does not install rabbitmq (because it's already
installed).

This enables us to continue using this role without forcing us to
upgrade rabbitmq.