A hidden secondaries upgrade script
This will connect to a cluster, find the hidden secondaries and upgrade
them to the current version specified in the mongo_3_0 play.

This required the creation of a new library module since our two current
ones return rs.status and call replSetReconfig, but neither return
rs.config, and rs.config is what tells you if a host is hidden and the
priority, which are needed for this.