Use force instead of setting fetch.prune
We originally started setting fetch.prune because it would cause the git
module to fail if there were refs locally that were not on the remote
(because they got deleted). Based on the Ansible devs' response on
https://github.com/ansible/ansible-modules-core/issues/637,
we can safely replace this task with setting the "force: true" flag
wherever we use the git module.