Support for Let's Encrypt
Implements automatic installation of SSL certificates for nginx using
Let's Encrypt certbot service.

Installs the certbot-auto and runs with list of domains specified in
LETS_ENCRYPT_DOMAINS. Validates with the Let's Encrypt server using
ACME http-01 (.well-known/acme) on a customisable port. This requires
an nginx server to be running.

Resulting certificates are symlinked to nginx certificate folder
and nginx configuration updated to match. To simplify the nginx
config-changing setup the ssl certificate configuration is moved
to a separate file for all dependent nginx sites.

Cronjob is installed to auto-update the certs in place on the
recommended twice-daily schedule.

Based on initial work by @tkeemon at:
https://github.com/tkeemon/configuration/tree/hackathon2016/lets-encrypt