Implement certbot, symlinking and copying as shell script
This alters the shell script generation for certbot-auto to also
handle the creating of symlinks, and copying of the config file
for nginx into place. The script is also moved to /opt. As a
result if the initial setup fails, the script can be re-run at
a later date to complete the installation.