Switch to manage_user from create_user
manage_user doesn't try to create a new user, but it doesn't handle
enrollments, so we have to enroll the users in a separate task using
another management command.

Additionally, it takes only password hashes, not passwords, so we have
to provide a pre-salted and pre-generated database string.

create_user used to autoconvert honor@example.com into honor for
username, we have to do that in the config now.