Save script and allow retry for getting certificate
Moves the logic for setting up the Let's Encrypt certificates
into a script. A semaphore file (/opt/.certbot-config-incomplete)
is used to mark the state of installation of certificates. If the
installation fails on first run, the script can be re-run.

Once installation is complete the semaphore is removed and the
script becomes a no-op.