Update to use the run_certbot script for the cron
As certbot-auto is smart enough to renew when running for
already requested domains, this update uses the same
run_cerbot script for the 2x daily cron. The script is reworked
to remove the semaphore.

The file-links and config copy is renundant (this could be moved
into an external file for the post-hook but not neccessary).
The script is also reworked to avoid unncessary nginx restart
(is in the post-hook).