Insights: Use py venv in bower install for sassc
We added a bower post-install task that pre-compiles the edX pattern library in
insights. This requires the executable sassc which is installed in
python-libsass. It was already being installed in a previous ansible task
(install application requirements), but we weren't sourcing the python
virtualenv where it was installed for the bower install task.