This was removed during the ansible 2.x upgrade.
Without this, 2 builds happen, PR and push.  While this is technically correct,
it doesn't work well with our current test logic that attepmts to test only things
that have changed.  The push branch runs all the tests which is not what we want.
We may want to enable this again in the future when we have jenkins running all the
docker builds in parallel.