Adds FORUM_MONGO_SSL_VERIFY
In case we're forced to use SSL with invalid (self signed) certs when
connecting to mongo the forum will fail to start because it won't be
able to connect to mongo. This implies using our forked version of
cs_comments_service for now (until we can get the path upstream).