Adds tag deploy:platform
This will ONLY checkout the latest code of the platform repo and restart
the required services/workers. It will not go and install new
requirements nor will it template the configuration files over and so
on.

It's to be used mostly for quick fixes where we know that the only thing
we want to do is deploy a few lines of new code.