Switch from git_add_course to standard import
git_add_course uses the platform's sysadmin dashboard which assumes it
can create an xlog mongo database and write into it without auth
configured.  This is not something we want to spend more time enabling,
so switch to our 'standard' import technique.

This relies on the demo role more than I would like, but we have a
backlog task to clean that role up, and can touch this at the same time
(this role has always imported a bunch of vars from demo, so I'm not
 adding any new deps).