matrix-auth was being overridden by Jenkins (down to 1.1)
This appears to be because it wasn't pinned and the shipped version
of the plugin won.  We have pinning code in jenkins_master that we
can use, but we have to copy down that list to add ours (ansible
doesn't allow "recursive references" of a variable passed back to
itself.

Continue process of trying to group plugins and their deps.