Updated LOG_DIR parameter for edxapp and xqueue
There was a typo on the `LOG_DIR` parameter for edxapp and xqueue
which results in the application trying to write logs to a directory
that never gets created, thereby silencing those logs. Updated to use
the `COMMON_LOG_DIR` value rather than composing the directory ad-hoc
in only those two places.