With the defined check, it's possible to have config syntax errors
Ansible will then skip the task and not write a config file.
I content that we want ansible to bail out if the dict it gets
in edx_service_config is invalid.

I'll need to change the two roles that pass no config (xqwatcher
and harstorage) to pass {} which will evaluate to false and skip
making a config.

Noticed this when manually testing config changes and saw missing
files rather than syntax errors.  In particular, this will make
our config checking code much better.