We have a customized version of core's hipchat
They made large changes in the callback code in 2.0->2.1
http://docs.ansible.com/ansible/latest/porting_guide_2.0.html#callback-plugins

I tested a few tweaks to make this run, but because it's the old v1
version, we're looking at a mostly complete rewrite.  Deleting it
and deciding if we actually valued this output was easier.

We've decided not to re-enable callbacks since all this output is in our
CI/CD and having it in hipchat isn't needed.