Retry the supervisor restart.
This is needed because there is a race condition between the stop and
start of supervisor.  The stop will not wait for all child process to
have shutdown by design because this behavior is needed for lms workers.

This means that when we try to do a stop and start together, and the first process
has not fully stopped, we get a failure in the startup of the second process.

There is no obvious way to deal with this in systemd so dealing with it here
instead.