Replaces EDXAPP_LMS_SPLIT_DOC_STORE_READ_PREFERENCE with EDXAPP_MONGO_READ_PREFERENCE
to indicate that this is the default read_preference for all mongo
replicasets.

Adds EDXAPP_LMS_DRAFT_DOC_STORE_READ_PREFERENCE, so it can differ from
the EDXAPP_MONGO_READ_PREFERENCE and default to PRIMARY.

Updates the CHANGELOG to indicate the variable changes, and to better
clarify their purpose and usage.