Switch to the nginx port used by xqueue (not the insights ports)
Override gunicorn to run on the nginx port, this is similar to what
other IDAs do, which is to use runserver and bind to the nginx port.
XQueue has multiple services (consumer and gunicorn) so we'll leave
docker-run.sh for now and just change the port specificed in the
supervisor file.

Bring the mysql/rabbit hostnames to the devstack standard names.