Have update_users only happen when services are running
Right now, it would run unless you are skip-tags'ing it (which is sort
of what we do in docker builds) or if the image has access to a database
(which is what happens during production builds).  It should instead be
explicitly requested.  I would remove it entirely, but we really want it
to run during sandbox builds, which use the ansible role.