Consolidate edxapp update assets scripts.
The asset compilation scripts edxapp-update-assets-lms and
edxapp-update-assets-cms duplicate a lot of work between them. Running
webpack twice also seems to cause an intermittent error where the
second build results in a different bundle hash than the first.

It is possible to run "paver update_assets" without specifying "lms" or
"cms", in which case it just compiles it all at the same time. This
removes the duplicate invocations of webpack and should speed the build
overall.