Fix config when analytics db name differs from config name
In https://github.com/edx/configuration/pull/4443/files I mistakenly changed
the ANALYTICS_DATABASE to the ANALYTICS_API_REPORTS_DB_NAME. In fact this
config refers to the database config identifer in the DATABASES dictionary.

This worked fine in sandboxes where both the config and the db are named
'reports', however when those configs differ, such as on staging, the health
check would fail as it tries to connect to the reporting database. This PR
fixes that change by setting this value back to 'reports'