Allow nginx to be templated but not enabled
If you set nginx_skip_enable_sites it will not symlink the items in
nginx_sites.  This means you have to make that change later.  EdX does
it with our pre_supervisor script, the same code that checks on
migrations being complete and also symlinks in supervisor processes to
run.

Remove the TODO: remove task for minos.
We stopped creating all the per-service dirs in edx_service