Moved Nginx log format name override (#4763)
When the `NGINX_LOG_FORMAT_NAME` var is set to something different than `p_combined` it breaks the Nginx configuration as there is no long a `p_combined` format defined for the main nginx config. This also prevents users from being able to specify their own format and using it throughout the various Nginx configs.

Removed the `NGINX_LOG_FORMAT_NAME` setting from the `log_format` definition to allow for the default log format to function properly. Users can define an alternate format with a different name and pass that via the `NGINX_LOG_FORMAT_NAME` to be used in the CMS/LMS etc.

Added log format customization to the xqueue Nginx template.