Adds a small plugin used by LabXchange.org to implement "pathways"
"Pathways" use the new Blockstore-based XBlock runtime to create small playlist-style learning experiences that consist of a few XBlocks in a short linear sequence. This plugin allows each "pathway" to freeze the version of each XBlock at the time that it gets added to the pathway, and to allow users to have different grades/state/completion status for an XBlock in a pathway vs. accessed directly from a content library.

The plugin only provides a Studio REST API which is consumed by the LabXchange backend and used for pathway CRUD operations. Otherwise, learners use the new runtime's existing XBlock REST API, which fully supports pathways (because pathways are implemented as a standard "learning context" plugin, just like a library or a course).

The plugin uses a simple security model where only an authorized service user (i.e. the LabXchange backend) can use the Studio REST API.