BOM-1241 - Add watchman to devstack.
Watchman is used by runserver in Django 2.2.  It will be installed by
default in newer versions of ubuntu but needs to be install manually for
ubuntu 16.04(xenial).

We choose to install it from a PPA instead of from source to not make
the build more complex.  The PPA is active and backed by a linux
news company.

# Conflicts:
#	playbooks/roles/common/tasks/main.yml