Remove create_user playbook, it breaks sshd config
In its current state this playbook will remove all other users from
AllowUsers in sshd_config, which will lock them out of the system. Just
disabling the part of the users role that updates sshd_config won't
solve this either because the added user needs to be added to
AllowUsers. The solution is to run the full users.yml playbook.