Fix ecomworker Docker image builds ARCHBOM-1270 (#5819)
This Docker image hasn't been built in 3 years, but is still used in the ecommerce-worker reop's Travis tests. Using this outdated image is now a blocker for merging Python 3.8 support, since it doesn't have Python 3.8 installed to test with (whereas the latest xenial-common images do). Updated the Dockerfile to work with the current tooling for image builds.