Removing override for credentials node version
this is a bit of a "kick the bucket" change, but I anticipate we
should not override the node version without overriding the npm
version as well. Those versions not matching appears to cause the
node-sass to be installed improperly, or for npm to not be installed
in an in-Docker usable way at all.

But either way this was prompted by the fact that this situation had
broken `make dev.static.credentials` in edx/devstack.