Validate certs, fix url for digicert ca download
This fixes a minor security issue where ssl certificate is not validated
for the digicert ca certificates download task.  It does this by using
a url that has a certificate that is able to be validated on a base
ubuntu image, and enables validate_certs on the ansible get_url task.

This is important, because without validating certs here, it's possible
for undesired ca certificates to be added to the root trust if this
playbook is run on an untrusted network.