Use a negated "is skipped" check for mariadb_new_debian_cnf
The "copy fetched file to other cluster members" task uses an "is
defined" test on the mariadb_new_debian_cnf fact to determine whether
the preceding "fetch debian.cnf file so start-stop will work properly"
task ran or not.

"is defined" is the wrong test to use here. As the Ansible
documentation[1] explains, if a task registers a variable, an "is
defined" test on a subsequent task will always return true, even if
the task has been skipped. Thus, we instead need to check with "is
skipped" to see if the task has been skipped (and in this case, we
want the reverse, so we negate the test).

[1] https://docs.ansible.com/ansible/latest/user_guide/playbooks_conditionals.html#conditions-based-on-registered-variables

(cherry picked from commit ede2669edeb0e5fdee1599b91afe508c98f1b8e6)