fix: allow mfe_deployer to run
The `mfe` role requires `COMMON_ECOMMERCE_BASE_URL` to be set, but that
variable is not defined in `common_vars`, `ecommerce`, or anywhere else.

Instead of introducing possibly breaking changes in  the `ecommerce` or
`mfe` roles, simply set it to the default `ECOMMERCE_ECOMMERCE_URL_ROOT`
here so that the native playbook can complete successfully.