fix: better decoding of ansible errors
Some failed log lines say "FAILED", but some don't, depending on whether
the task is loop or not.  Sometimes an earlier FAILED message is
actually ignored, but is found by the old check.

Look for the leading "failed:" or "fatal:" lable instead to always get
the actual error that stopped the playbook.