feat: add celerybeat configuration
The current implementation of celerybeat usage works properly if and
only if one concurrent worker process runs per queue using the nested
beat process started by the `--beat` flag. In case the the celerybeat
process is running as a separate process not as part of the workers,
that partially solves the problem. Either in that case, if multiple
instances are running in a cluster, without proper process supervising
across instances will result in (at least) dumplicated scheduling,
meaning that the scheduler will cause more work for the workers and
results will be duplicated.

Taking the above into consideration, to solve this issue, this commit
introduces single-beat package that wraps the celerybeat process and
keeps track of the process that acquired the lock in redis. In case one
of the instances get killed or the locking process crashes, the lock

refactor: use boolean flag to enable or disable celery beat
Co-authored-by: Joseph Mulloy <jdmulloy@users.noreply.github.com>

refactor: use boolean flag to enable or disable celery beat
Co-authored-by: Joseph Mulloy <jdmulloy@users.noreply.github.com>

Unindent a few entries that needd to be unindented (#6450) (#6467)
(cherry picked from commit 945fe9df4aaef4bc1dfc565eab4bbcc0b12f9ee8)
Co-authored-by: Chris Pappas <christopappas@users.noreply.github.com>